Add color channel masking to RGBA's hex conversion
The color_rgba object's to_hex() function, which converts the RGBA
values to a hex string, now accepts a bit mask, with which the user can
control which color channels are included in the hex string. (See
further comments in code.)

Note that the to_hex() function by default now returns a hex string
which includes the alpha channel. It used to return #rrggbb, i.e. no
alpha.