Update README.md
Update perf test results to reflect recent optimization to textured
rendering.

Note that results previously marked to be for the Nexus 9 are actually for
the Huawei MediaPad T1-A21L. The tablet's not mine, so I wasn't clear on
which make or model it was, but that's been clarified now.