Add integration test for render output
It's a page that renders a bunch of images and compares them against
reference ones to see whether there are any differences between the
rendering and the corresponding reference. None should exist. The images
can also be compared visually by looking at them.

Canvas rendering can produce slight variations depending on the
underlying hardware and operating environment. So the automatic test
may fail if these slight variations are produced relative to the
reference images, even when there isn't actually anything wrong with
the rendering per se. In that case, you might generate a set of new
reference images for that platform.