Add support for auxiliary render buffers
More in-depth documentation on auxiliary buffers to follow, but in brief,
they're buffers that let you tell the renderer to provide additional,
customized metadata for each pixel rendered. For instance, you might
attach an id value to each n-gon, and tell the renderer to store for
each pixel rendered the id of the source n-gon in an auxiliary buffer;
so you can do mouse-picking after rendering.