Use persistent cache for transformed n-gons
All n-gons transformed for a given frame are now placed in a permanent
cache from which they'll be accessed for rasterization. (The n-gons in the
cache are overwritten by new ones the next frame, but the cache container
remains across frames).

This reduces the load on garbage collection (unless the "painter" depth
sorting mode is used, which still results in extra temporary objects
generated to accommodate the sorting).