ngon-fill: Remove requirement for verts to be arranged in a loop
When sorting an n-gon's vertices between the left and right sides, we
don't really need to sort the left side as ascending on Y and the right
as descending. We can just sort them both ascending, which saves us one
extra sorting of the right-side vertex list into descending.

This commit also does some work to clarify the in-code comments for the
n-gon drawing algo.