ngon-fill: Remove triangle-specific vertex sorting code path
Although intended only as a placeholder implementation in the first
place, it's taking up a bunch of space in the code without contributing
much back at the moment. I think it should be removed for now, and
re-added in the future if a specific need for it arises (e.g. it can be
implemented tangibly more efficiently than the generic n-gon path).