ngon-fill: Offset vertical flip for ortho-mapped textures
This causes the render integration test for ortho mapping to fail; but on
the other hand is required for RallySportED-js to render properly.

I'll look into this bug more carefully at some point, but for now, let's
ignore what the integration test says and go with what the program needs.