Add toggle to allow/disallow the use of window.alert()
Previously, window.alert() was used by the renderer to directly alert
the user to an error being thrown. But in production environments, this
can be disruptive to the user experience, so a toggle is added by which
to set whether the renderer is allowed to do this.

(That said, the renderer is currently in alpha and so its use in
production environments is not necessarily encouraged.)