Add groundwork for pixel shaders
This commit adds a new render buffer - the fragment buffer - which
contains for each rendered pixel metadata about the state of the
renderer at that pixel; things like interpolated texture and world
coordinates. A shader function can then make use of this information to
apply custom effects to the rendered image.