samples/oblique-2d-tiles: Shade tiles in unit size
Instead of computing a shade value for each of a tile's corners, we do
it for the tile's unit-size base square. This effectively flattens all
tiles (e.g. vertical walls) to the height of ground tiles, creating a
more uniform shading across all tiles.

Whether this is a better lighting strategy overall depends on the
tilemap's topology (e.g. whether there are many stacked layers or tall
walls), but it works better for this particular render sample.