rasterizer: Use Uint32Array to access pixel buffer
Can be 5-15% faster, depending on the scene and resolution. The downside
is it's not compatible as-is with big-endian systems - but I don't know
that this would be a major problem.

NOTE: This commit removes Phong shading. It had never been fully
implemented, and I don't think it fits in. Maybe it'll be re-added in
the future, but for now let's get rid of it.