Implement scaffolding for paletted rendering
This commit includes a paletted implementation of <canvas> as well as
preliminary code for rendering into it. The canvas uses an 8-bit pixel
buffer, allowing the rasterizer to write only palette indices instead of
full 32-bit color values.

The implementation is work in progress and not ready for use outside
of testing.

Some code refactoring was also done.