base-modules/rasterize: Move raster paths into their own source files
They're now tentatively called raster shaders. A raster shader receives
the edges of a polygon and some other related data, and is responsible
for rasterizing the polygon into the pixel buffer.

In the future, the end user might be able to add custom raster shaders,
but I haven't yet worked out an architecture for it.