api/render: Implement support for multiple render states
A render state identifier can now be provided to render() via
options.state, determining which render state is to be used by that
render call.

This deprecates the feature of the previous commit that enabled custom
pixel buffers, as the render state encompasses among other things the
pixel buffer. In other words, using a second render state provides a
second pixel buffer.