Internally, use import instead of global Rngon object
This introduces some changes to the public API that the README's API
reference hasn't yet been updated with.

Note: 'Rngon.state.active' is now 'Rngon.state.default'.

Note: This commit removes support for async rendering. Ideally it would
be done in a separate commit, but lumping it into this commit was more
convenient for me. I might re-enable async support, but so far I'm not
sure it's a useful feature, and it's extra work to maintain.