api/state: Add a vertex normal cache
The existing ngon and vertex caches helped reduce GC overhead, so why
not add a cache for vertex normals as well.

That said, at least in Google Chrome, this new cache doesn't actually
seem to reduce GC (nor increase it). But let's just leave it in for now
and look into it in more detail at some point.